+++ scanpipe/pipes/scancode.py
@@ -324,10 +324,11 @@ def run_scancode(location, output_file, options, raise_on_error=False):
         *options_from_settings,
         *options,
         f"--processes {max_workers}",
+        "--verbose",
         f"--json-pp {shlex.quote(output_file)}",
     ]
 
-    exitcode, output = pipes.run_command(scancode_args)
+    exitcode, output = pipes.run_command(scancode_args, log_output=True)
     if exitcode > 0 and raise_on_error:
         raise ScancodeError(output)