+++ scanner/tests/data/reference_scan/code/zlib/zutil.h
@@ -1,4 +1,0 @@
-/* zutil.h -- internal interface and configuration of the compression library
- * Copyright (C) 1995-2011 Mark Adler
- * For conditions of distribution and use, see copyright notice in zlib.h
- */