+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 // Release notes
 // -------------
 
+### Unreleased
+
+- Log the outputs of run_scancode as progress indication.
+  https://github.com/nexB/scancode.io/issues/300
+
 ### v21.8.2
 
 - Upgrade ScanCode-toolkit to version 21.7.30